Add Basic Working Client & Server
Add a basic client that requests github login details, posts them to the
server, fetches your repos & their traffic, then renders it in a table
for you.

The server makes the Repository & Traffic API requests to github for us.
This is necessary because we cannot make BasicAuth XHR requests to
domains with wildcard Access-Control-Allowed-Origins(which includes
github).

This means that entered passwords are transmitted to the server in
cleartext and that this application shouldn't be hosted publicly!